+++ src/components/Search/index.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import Wrapper from "../Container/Wrapper";
-import FilterModal from "../../modal/FilterModal";
+import FilterModal from "../modal/FilterModal";
 
 export default function Search() {
   return (
